refactor(converters): rename `_helpers.py` to `_utils.py`
