+++ pydantic_jsonschema/converters/_discriminator.py
@@ -16,7 +16,7 @@ from pydantic.experimental.missing_sentinel import MISSING
 
 from pydantic_jsonschema.types import Reference, Schema
 
-from ._helpers import unwrap
+from ._utils import unwrap
 
 __all__ = ["discriminator_property"]
