+++ pydantic_jsonschema/converters/_object_keywords.py
@@ -18,7 +18,7 @@ from pydantic.experimental.missing_sentinel import MISSING
 
 from pydantic_jsonschema.types import Schema
 
-from ._helpers import unwrap
+from ._utils import unwrap
 
 __all__ = [
     "build_dependent_required",
